prep-populate song preview on edit page
